Added stuff based on scp's meta-stlinux-20140207 package
SCP does provide his stuff at http://forum.mbremer.de/ With the permission
from scp I do provide a copy here.

!!! Attention !!!
This is untested stuff

+++ recipes-core/glibc/glibc-locale_2.10.bb
@@ -36,27 +36,28 @@ PROVIDES = "virtual/libc-locale${PKGSUFFIX}"
 PACKAGES = "localedef${PKGSUFFIX} ${PN}-dbg"
 
 PACKAGES_DYNAMIC = "locale-base-* \
-                    glibc-gconv-*${PKGSUFFIX}  glibc-charmap-*  glibc-localedata-*  glibc-binary-localedata-*"
+                    eglibc-gconv-*${PKGSUFFIX} eglibc-charmap-*${PKGSUFFIX} eglibc-localedata-* eglibc-binary-localedata-* \
+                    glibc-gconv-*${PKGSUFFIX}  glibc-charmap-*${PKGSUFFIX}  glibc-localedata-*  glibc-binary-localedata-*"
 
 # Create a glibc-binaries package
-ALLOW_EMPTY_${BPN}-binaries = "1"
-PACKAGES += "${BPN}-binaries"
-RRECOMMENDS_${BPN}-binaries =  "${@" ".join([p for p in d.getVar('PACKAGES', True).split() if p.find("glibc-binary") != -1])}"
+ALLOW_EMPTY_${BPN}-binaries${PKGSUFFIX} = "1"
+PACKAGES += "${BPN}-binaries${PKGSUFFIX}"
+RRECOMMENDS_${BPN}-binaries${PKGSUFFIX} =  "${@" ".join([p for p in d.getVar('PACKAGES', True).split() if p.find("glibc-binary${PKGSUFFIX}") != -1])}"
 
 # Create a glibc-charmaps package
-ALLOW_EMPTY_${BPN}-charmaps = "1"
-PACKAGES += "${BPN}-charmaps"
-RRECOMMENDS_${BPN}-charmaps =  "${@" ".join([p for p in d.getVar('PACKAGES', True).split() if p.find("glibc-charmap") != -1])}"
+ALLOW_EMPTY_${BPN}-charmaps${PKGSUFFIX} = "1"
+PACKAGES += "${BPN}-charmaps${PKGSUFFIX}"
+RRECOMMENDS_${BPN}-charmaps${PKGSUFFIX} =  "${@" ".join([p for p in d.getVar('PACKAGES', True).split() if p.find("glibc-charmap${PKGSUFFIX}") != -1])}"
 
 # Create a glibc-gconvs package
-ALLOW_EMPTY_${BPN}-gconvs = "1"
-PACKAGES += "${BPN}-gconvs"
-RRECOMMENDS_${BPN}-gconvs =  "${@" ".join([p for p in d.getVar('PACKAGES', True).split() if p.find("glibc-gconv") != -1])}"
+ALLOW_EMPTY_${BPN}-gconvs${PKGSUFFIX} = "1"
+PACKAGES += "${BPN}-gconvs${PKGSUFFIX}"
+RRECOMMENDS_${BPN}-gconvs${PKGSUFFIX} =  "${@" ".join([p for p in d.getVar('PACKAGES', True).split() if p.find("glibc-gconv${PKGSUFFIX}") != -1])}"
 
 # Create a glibc-localedatas package
-ALLOW_EMPTY_${BPN}-localedatas = "1"
-PACKAGES += "${BPN}-localedatas"
-RRECOMMENDS_${BPN}-localedatas =  "${@" ".join([p for p in d.getVar('PACKAGES', True).split() if p.find("glibc-localedata") != -1])}"
+ALLOW_EMPTY_${BPN}-localedatas${PKGSUFFIX} = "1"
+PACKAGES += "${BPN}-localedatas${PKGSUFFIX}"
+RRECOMMENDS_${BPN}-localedatas${PKGSUFFIX} =  "${@" ".join([p for p in d.getVar('PACKAGES', True).split() if p.find("glibc-localedata${PKGSUFFIX}") != -1])}"
 
 DESCRIPTION_localedef = "glibc: compile locale definition files"
 
@@ -67,15 +68,25 @@ FILES_${MLPREFIX}glibc-gconv = "${libdir}/gconv/*"
 FILES_${PN}-dbg += "${libdir}/gconv/.debug/*"
 FILES_localedef${PKGSUFFIX} = "${bindir}/localedef"
 
-LOCALETREESRC = "${STAGING_INCDIR}/glibc-locale-internal-${MULTIMACH_TARGET_SYS}"
+LOCALETREESRC = "${STAGING_INCDIR}/eglibc-locale-internal-${MULTIMACH_TARGET_SYS}"
 
 do_install () {
 	mkdir -p ${D}${bindir} ${D}${datadir} ${D}${libdir}
-	cp -fpPR ${LOCALETREESRC}/${bindir}/* ${D}${bindir}
-	cp -fpPR ${LOCALETREESRC}/${libdir}/locale ${D}${libdir}
-	cp -fpPR ${LOCALETREESRC}/${libdir}/gconv ${D}${libdir}
-	cp -fpPR ${LOCALETREESRC}/${datadir}/i18n ${D}${datadir}
-	cp -fpPR ${LOCALETREESRC}/${datadir}/locale ${D}${datadir}
+	if [ -n "$(ls ${LOCALETREESRC}/${bindir})" ]; then
+		cp -fpPR ${LOCALETREESRC}/${bindir}/* ${D}${bindir}
+	fi
+	if [ -e ${LOCALETREESRC}/${libdir}/locale ]; then
+		cp -fpPR ${LOCALETREESRC}/${libdir}/locale ${D}${libdir}
+	fi
+	if [ -e ${LOCALETREESRC}/${libdir}/gconv ]; then
+		cp -fpPR ${LOCALETREESRC}/${libdir}/gconv ${D}${libdir}
+	fi
+	if [ -e ${LOCALETREESRC}/${datadir}/i18n ]; then
+		cp -fpPR ${LOCALETREESRC}/${datadir}/i18n ${D}${datadir}
+	fi
+	if [ -e ${LOCALETREESRC}/${datadir}/locale ]; then
+		cp -fpPR ${LOCALETREESRC}/${datadir}/locale ${D}${datadir}
+	fi
 	cp -fpPR ${LOCALETREESRC}/SUPPORTED ${WORKDIR}
 }
